+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog
  * removed argument `$field` from `key()`
  * removed argument `$strict` from `true()`, `false()` and `oneOf()`
  * added `atLeastOne()` selector
+ * added `exactly()` selector
 
 * 1.0.0-beta (2015-03-19)